requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==5.0.2
django-binary-database-files==1.0.10
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.0.0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.1.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
